Move command.ts and commands.ts to the commands directory
These command implementations are a specific implementation of handling interaction,
so it should not be in the global folder.
