+++ api/src/discord/commands/command.ts
