+++ api/src/discord/commands/workingOn.ts
@@ -1,5 +1,5 @@
 import { ApplicationCommandType, Client, CommandInteraction } from "discord.js";
-import { Command } from "../command";
+import { Command } from "./command";
 import {
   userStartsWorkingOnTask,
   userStopsWorkingOnTask,
