+++ api/src/discord/listeners/ready.ts
@@ -1,6 +1,6 @@
 import { ActivityType, Client } from "discord.js";
 
-import { Commands } from "../commands";
+import { Commands } from "../commands/commands";
 import interactionCreate from "./interactionCreate";
 import fs from "fs";
 import config from "../../config";
